v4 defconfig: increase NuttX RX & TX UART buffer sizes for TELEM2
The buffer sizes were too small for a companion with log streaming.
Mavlink showed tx errors, which are gone with this patch.
The TX rate was around 50KB/s in the tests.
The RX rate was around 10KB/s, not sure if there were drops there as well.

TODO: this is definitely too high, check how much we really need.