Simple mixer: add output slew rate
This adds the option to limit the rate of change (slew rate) of an output that's mixed by a simple mixer.
For it to work, the param MIX_SIMP_SLEW has to be set to true, and a non-zero slewrate must
be specified in the mixer file (by adding number at end of output scaler O:, at the 6th position).

Signed-off-by: Silvan Fuhrer <silvan@auterion.com>