navigator: always accept yaw after takeoff
Navigator sets a 3D waypoint for takeoffs that includes a heading
setpoint. When the yaw / heading is reset during takeoff, this
yaw setpoint is outdated and wrong. For this reason FlightTaskAuto
is currently not even trying to track yaw during takeoff.

This is just a temporary solution. The proper fix is to modify
Navigator such that a heading reset is handled correctly and the
waypoint and resulting setpoint is updated.