commander: always check ekf2 in preflight check
I don't understand why we should wait 10 seconds until checking ekf2. It
makes sense not to complain to the user in that time, but we should
probably still check it already.

This commit should fix the case where MAVSDK tries to arm too early
(within 10 seconds) and arming is denied silently.