Check SD card missing only once and store the result
statfs accesses the file-system and can block for seconds. The SD card check is part of the preflight checks and blocks the high priority thread in commander which can cause various issues. The SD card is mounted during boot and the state will anyways not change after the check is executed once.