commander: alternative for silent ekf2 check
This is an alternative fix for the ekf2 check. With the new logic, we
don't run the ekf2 check within the first 10 seconds, unless
specifically requested by e.g. arming at which point report_failures is
set. In that case we do run the check and display whatever comes up.