MavlinkMissionManager: Mute "Operation timeout" error message
The operation timeout occurs from time to time during boot. MISSION_REQUEST_LIST can timeout because of the SD card blocking or delays in the MAVLink communication. If the mission count is greater than zero, there is another error message in AMC that the download failed. There is no message in AMC when mission count is zero but then the operation timeout is not critical and not actionable for the user.
The logging is changed from a critical MAVLink message to PX4_WARN.