arx_rls: make sure the buffer is filled before starting fitting
The RLS is really sensitive to the first few updates as the covariance
is initially really large. Updating it with a buffer (history of input
and output values) partially empty (i.e.:full of zeros) can make it
converge to a bad estimate if the first values are far from zero.