SysId: low-pass filter the y signal
By low-passing the output signal with the same filer used for the input,
the system to identify now includes the gyro filter that is part of the
model (looking at the controller point-of-view). This makes more sense
as the sytem to be controlled includes the gyro filter.
Since the filtering is now symmetric, the identification is less
sensitive to filter initialization and reset