Fixed-wing Position controller: introduce Eco, Normal and Dash cruise modes
Eco: have less tight altitude control and reducd max throttle and climbrate
Dash: increase airspeed setpoint

Eco can be automatically started if within some altitude band, and disabled
if that is breched.
Dash can be started via a param, and reset to Normal after 60s.
Instead of relying on params to start Dash or disable Eco, we want to use
mavlink commands from the groundstation as a next step.

Signed-off-by: Silvan Fuhrer <silvan@auterion.com>