commander: fix battery failsafe without RTL
We trigger battery failsafe actions earlier if we are far away from home
and RTL would take longer. This makes sense when our failsafe action is
to do RTL, however, when the action is set to just land or warn only, it
doesn't make much sense to trigger the failsafe earlier.

Therefore, with this change, we're ignoring the RTL estimate if RTL is
not configured as a battery failsafe action.