mavlink: fix runaway case during mission upload
When trying to upload missions using MAVDSK we ran into the case where
more and more retries happened and eventually it would time out.

This could be traced to the case where previous (old) mission item
messages would trigger re-requesting of the current mission item even if
it is already on the way.
This case can be avoided by just ignoring old mission items.

I believe the change matches the spec which says that the drone is
responsible of requesting mission items, as well as checking for
timeouts.
Also see: https://mavlink.io/en/services/mission.html