Adding new px4 signing tool, to support secure BTL.
A replacement for the current cryptotool.py. This tool can sign .px4 and .bin files with keys provided from command line.
It rewrites the TOC within the px4 bin according the new keys. It only supports one TOC entry. And it finally applies the signature at the end of the binary.
Fixing allignment with new binary.