custom_sitl_test: reduce speed factor
The logs show, that the custom sitl tests are currently running at
an effective speed-factor of 1.3 or something. And occasionally
MAVSDK loses connection during the tests, causing a failure.

I have observed something similar on the local machine when the
speed-factor is ridiculously high. So let's test if this brings
some improvement.