modeCheck: allow arming in init mode
to not break all the existing MAVSDK sitl tests/use cases.

This passes upstream because the autopilot initializes in manual mode
which is an allowed mode to arm in (even without having rc).
While init mode for good reason was not allowed for takeoff.
But for the tests it has to be because the MAVSDK tests first arm
(in init mode) and then switch to the mode that they want to
run in e.g. mission. I think the entire MAVSDK worflow works
like that but that has to be verified and discussed.

Arming in Init mode was obviously also possible before and I
checked currently it's not possible to take off in init mode because
it executes the vehicle_rates_setpoint that was never published before
and initialized to zero because Init mode runs first and can never be
switched to again.